added exception to delete method

+++ api/dynamodb/dynamodb.py
@@ -32,10 +32,13 @@ class DynamodbDao:
         return response["Items"]
 
     def delete_item(self, city, attraction):
-        self.dynamodb_table.delete_item(
+        if self.get_attraction(city, attraction) is None:
+            return None
+
+        response = self.dynamodb_table.delete_item(
             Key={
                 "PK": city,
                 "SK": attraction,
             }
         )
-        return "Deleted"
+        return response
